+++ modulo2/projeto-labefy/projeto-labefy/src/components/AddPlaylistCard/style.js
@@ -1,9 +1,6 @@
-import React, { Component } from 'react'
 import styled from "styled-components";
 
-import Add from '../assets/addition.png'
-
-const Container = styled.div`
+export const Container = styled.div`
   background-color: #DA0037;
   border-radius: 8px;
   display: flex;
@@ -11,7 +8,7 @@ const Container = styled.div`
   align-items: center;
   justify-content: center;
 `
-const Button = styled.button`
+export const Button = styled.button`
   height: 50%;
   margin: 0 10px 10px 10px;
   border-radius: 7px;
@@ -25,14 +22,14 @@ const Button = styled.button`
 
     }
 `
-const Img = styled.img`
+export const Img = styled.img`
   width: 60px;
 `
-const Span = styled.span`
+export const Span = styled.span`
   font-size: 30px;
   margin: 10px;
 `
-const Input = styled.input`
+export const Input = styled.input`
   width: 75%;
   height: 30px;
   margin-bottom: 5px;
@@ -40,18 +37,3 @@ const Input = styled.input`
   color: black;
   padding: 5px;
 `
-
-export default class AddPlaylistCard extends Component {
-  render() {
-    return (
-      <Container>
-        <Button>
-          <Img src={Add} />
-          <Span>Add Playlist</Span>
-        </Button>
-
-        <Input type="text" placeholder='> Digite aqui o nome da sua Playlist' />
-      </Container>
-    )
-  }
-}
